requirements.txt
Django==4.1.5
django-cors-headers==3.13.0
django-crispy-forms==1.14.0
djangorestframework==3.14.0
python-decouple==3.7
requests==2.28.2
psycopg2-binary==2.9.7